+++ NEWS.md
@@ -1,5 +1,6 @@
 # teal.slice 0.2.0.9007
 
+* `ui/srv_filter_panel` module can be called simultaneously in multiple places within app and still refer to the single `FilteredData` object.
 * Examples now use `scda.2022` instead of `scda.2021`.
 * Transferred data hashing step in `FilteredDataset` to `teal`.